# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/waybills": {
      "get": {
        "operationId": "WaybillsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      },
      "post": {
        "operationId": "WaybillsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
    "/api/v1/waybills/by-number/{number}": {
      "get": {
        "operationId": "WaybillsController_getByNumber",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
    "/api/v1/waybills/{id}": {
      "get": {
        "operationId": "WaybillsController_getById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
    "/api/v1/waybills/{id}/status": {
      "post": {
        "operationId": "WaybillsController_updateStatus",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
    "/api/v1/qr/resolve": {
      "post": {
        "operationId": "WaybillsController_resolveQr",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
    "/api/v1/waybills/{id}/accept": {
      "post": {
        "operationId": "WaybillsController_accept",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
...
    "/api/v1/reports": {
      "get": {
        "operationId": "ReportsController_report",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/export": {
      "get": {
        "operationId": "ReportsController_export",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/transfers": {
      "get": {
        "operationId": "TransfersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "transfers"
        ]
      },
      "post": {
        "operationId": "TransfersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/v1/transfers/{id}/confirm": {
      "post": {
        "operationId": "TransfersController_confirm",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/v1/transfers/{id}/cancel": {
      "post": {
        "operationId": "TransfersController_cancel",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "transfers"
        ]
      }
    },
...
  },
  "info": {
    "title": "BIOFLOW API",
    "description": "Control of extraction, transport, acceptance, inventory, washing and production.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}